requirements.txt
google-cloud-vision
jupyter
pillow